requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pg8000==1.30.5
python-dotenv==1.0.1
faker==23.2.1